move a bunch of things
