+++ src/main/java/net/cytonic/cytosis/commands/CommandHandler.java
@@ -5,10 +5,7 @@ import net.cytonic.cytosis.commands.defaultMinecraft.GamemodeCommand;
 import net.cytonic.cytosis.commands.defaultMinecraft.TeleportCommand;
 import net.cytonic.cytosis.commands.friends.AddFriendCommand;
 import net.cytonic.cytosis.commands.moderation.BanCommand;
-import net.cytonic.cytosis.commands.server.CreateInstanceCommand;
-import net.cytonic.cytosis.commands.server.PodDetailsCommand;
-import net.cytonic.cytosis.commands.server.ShutdownInstancesCommand;
-import net.cytonic.cytosis.commands.server.StopCommand;
+import net.cytonic.cytosis.commands.server.*;
 import net.minestom.server.command.CommandManager;
 import net.minestom.server.entity.Player;
