+++ src/main/java/net/cytonic/cytosis/commands/server/TimeCommand.java
@@ -1,4 +1,4 @@
-package net.cytonic.cytosis.commands;
+package net.cytonic.cytosis.commands.server;
 
 import net.cytonic.cytosis.Cytosis;
 import net.minestom.server.command.builder.Command;
@@ -41,32 +41,26 @@ public class TimeCommand extends Command {
                     case "day" -> {
                         Cytosis.getDefaultInstance().setTime(1000); // Day
                         sender.sendMessage(MM."<GREEN>Time set to day.");
-                        return;
                     }
                     case "night" -> {
                         Cytosis.getDefaultInstance().setTime(13000); // Night
                         sender.sendMessage(MM."<GREEN>Time set to night.");
-                        return;
                     }
                     case "midnight" -> {
                         Cytosis.getDefaultInstance().setTime(18000); // Midnight
                         sender.sendMessage(MM."<GREEN>Time set to midnight.");
-                        return;
                     }
                     case "noon" -> {
                         Cytosis.getDefaultInstance().setTime(6000); // Noon
                         sender.sendMessage(MM."<GREEN>Time set to noon.");
-                        return;
                     }
                     case "sunrise" -> {
                         Cytosis.getDefaultInstance().setTime(23000); // Sunrise
                         sender.sendMessage(MM."<GREEN>Time set to sunrise.");
-                        return;
                     }
                     case "sunset" -> {
                         Cytosis.getDefaultInstance().setTime(12000); // Sunset
                         sender.sendMessage(MM."<GREEN>Time set to sunset.");
-                        return;
                     }
                 }
             }
