+++ src/main/java/net/cytonic/cytosis/commands/server/VersionCommand.java
@@ -1,4 +1,4 @@
-package net.cytonic.cytosis.commands;
+package net.cytonic.cytosis.commands.server;
 
 import net.cytonic.cytosis.Cytosis;
 import net.minestom.server.command.builder.Command;
